Extract shared git repo test helpers
