+++ test/test_gem_version_tag_agreement.rb
@@ -3,7 +3,6 @@
 require "test_helper"
 require "rake"
 require "stringio"
-require "tempfile"
 require "tmpdir"
 
 # Bundler loads the gemspec once, when bundler/gem_tasks is required, which is
@@ -12,6 +11,8 @@ require "tmpdir"
 # confirmation messages come from the copy bundler is still holding. Left alone
 # they name the version before the bump.
 class TestGemVersionTagAgreement < Minitest::Test
+  include GitRepoHelpers
+
   def setup
     @rake = Rake::Application.new
     Rake.application = @rake
@@ -47,9 +48,7 @@ class TestGemVersionTagAgreement < Minitest::Test
   private
 
   def setup_project(version:)
-    system("git init", out: File::NULL, err: File::NULL)
-    system("git config user.name 'Test'", out: File::NULL, err: File::NULL)
-    system("git config user.email 'test@example.com'", out: File::NULL, err: File::NULL)
+    init_git_repo
 
     File.write("version.rb", "VERSION = \"#{version}\"\nRELEASE_DATE = \"Unreleased\"\n")
     File.write("CHANGELOG.md", <<~CHANGELOG)
@@ -76,21 +75,8 @@ class TestGemVersionTagAgreement < Minitest::Test
       end
     RUBY
 
-    system("git add .", out: File::NULL, err: File::NULL)
-    system("git commit -m 'Initial'", out: File::NULL, err: File::NULL)
-    system("git tag v#{version}", out: File::NULL, err: File::NULL)
-  end
-
-  def create_commit_with_trailer(subject, trailer)
-    filename = "test_#{Time.now.to_f}.txt"
-    File.write(filename, "content")
-    system("git add #{filename}", out: File::NULL, err: File::NULL)
-
-    Tempfile.create("commit_msg") do |f|
-      f.write("#{subject}\n\n#{trailer}")
-      f.flush
-      system("git commit -F #{f.path}", out: File::NULL, err: File::NULL)
-    end
+    commit_everything("Initial")
+    tag_version(version)
   end
 
   def create_rakefile
